requirements.txt
mysql-connector-python==8.0.29
python_telegram_bot==20.0a1
telegram==0.0.1
python-dotenv==0.21.0
